+++ Playground/Playground/AppSetup.cs
@@ -4,11 +4,11 @@ using Playground.Features.Animation;
 using Playground.Features.BattleTest;
 using Playground.Features.CssPreviewer;
 using Playground.Features.Editor;
-using Playground.Features.Editor.Services;
 using Playground.Features.Gallery;
 using Playground.Features.Gallery.Services;
 using Playground.Features.Home;
 using Playground.Features.Linear;
+using Playground.Features.Share;
 using SimpleInjector;
 using Xamarin.Forms;
 using GradientEditorPage = Playground.Features.Editor.GradientEditorPage;
