+++ Playground/Playground/Features/Share/ExportData.cs
@@ -1,0 +1,11 @@
+﻿using MagicGradients;
+
+namespace Playground.Features.Share
+{
+    public class ExportData
+    {
+        public IGradientSource GradientSource { get; set; }
+        public Dimensions GradientSize { get; set; }
+        public BackgroundRepeat GradientRepeat { get; set; }
+    }
+}
