+++ Playground/Playground/Features/Share/IGradientExporter.cs
@@ -1,0 +1,8 @@
+﻿namespace Playground.Features.Share
+{
+    public interface IGradientExporter
+    {
+        string ExportCss(ExportData data);
+        string ExportRaw(ExportData data);
+    }
+}
